backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.2.10=pypi_0
backcall=0.2.0=pypi_0
bcrypt=3.2.0=pypi_0
branca=0.3.1=pypi_0
certifi=2019.11.28=pypi_0
cffi=1.14.3=pypi_0
chardet=3.0.4=pypi_0
colorama=0.4.3=pypi_0
cryptography=2.9.2=pypi_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
django=2.2.7=pypi_0
djangorestframework=3.10.3=pypi_0
folium=0.10.1=pypi_0
idna=2.8=pypi_0
ipython=7.16.1=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.17.2=pypi_0
jinja2=2.10.3=pypi_0
joblib=0.14.0=pypi_0
jwt=1.0.0=pypi_0
kiwisolver=1.1.0=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.1.2=pypi_0
mysqlclient=2.0.1=pypi_0
numpy=1.17.3=pypi_0
pandas=0.25.3=pypi_0
parso=0.7.1=pypi_0
pickleshare=0.7.5=pypi_0
pip=20.2.2=pypi_0
prompt-toolkit=3.0.7=pypi_0
pycparser=2.20=pypi_0
pygments=2.6.1=pypi_0
pyparsing=2.4.5=pypi_0
python=3.6.8=h9f7ef89_7
python-dateutil=2.8.1=pypi_0
pytz=2019.3=pypi_0
requests=2.22.0=pypi_0
setuptools=49.6.0=pypi_0
six=1.13.0=pypi_0
sqlite=3.33.0=h2a8f88b_0
sqlparse=0.3.0=pypi_0
traitlets=4.3.3=pypi_0
urllib3=1.25.7=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wcwidth=0.2.5=pypi_0
wheel=0.35.1=py_0
wincertstore=0.2=py36h7fe50ca_0
zlib=1.2.11=h62dcd97_4
